+++ apps/pm_mobile/lib/src/api/lease_api.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'lease_api.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$leaseApiHash() => r'53aad1a06d82716deb9b9ce67a68728169bb7c4f';
+
+/// See also [leaseApi].
+@ProviderFor(leaseApi)
+final leaseApiProvider = AutoDisposeProvider<LeaseApi>.internal(
+  leaseApi,
+  name: r'leaseApiProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$leaseApiHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef LeaseApiRef = AutoDisposeProviderRef<LeaseApi>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member